+++ src/main/java/search/SearchByPlayerAttributes.java
@@ -35,9 +35,9 @@ public class SearchByPlayerAttributes {
 
     public static boolean addPlayer(Map<String, Number> attributes, Map<String, Pair<Number, Number>> queries){
         for (Map.Entry<String, Pair<Number, Number>> entry: queries.entrySet()){
-            double val = (double) attributes.get(entry.getKey());
-            double min = (double) entry.getValue().getValue0();
-            double max = (double) entry.getValue().getValue1();
+            int val = (int) attributes.get(entry.getKey());
+            int min = (int) entry.getValue().getValue0();
+            int max = (int) entry.getValue().getValue1();
             if (!((min <= val) & (val <= max))){ return false;}
         }
       return true;
